disable buttons on loading, protect against double clicking that calls asyncronous requests, show spinner on long requests

+++ src/main/webapp/WEB-INF/js/controllers/schedule.js
@@ -20,19 +20,20 @@ angular.module('sequoiaGroveApp')
         $translate) {
 
 /************** Login Redirect, Containers and UI settings **************/
+  $rootScope.lastPath = '/schedule';
+  $scope.saving = false;
+  $scope.importing = false;
 
   // user is not logged in
   if ($rootScope.loggedIn == false) {
     $location.path('/login');
   }
-  $rootScope.lastPath = '/schedule';
 
   $scope.activeTab = 'schedule';
   $scope.selectedId = 0;
   $scope.newDelivery = '';
   $scope.selectedPid = 0;
   $scope.empEditSearch = '';
-  $scope.saving = false;
   $scope.selectedShift = {
     idx : -1,
     sid : -1,
@@ -295,7 +296,7 @@ angular.module('sequoiaGroveApp')
       $scope.updateShifts.push(paramObj);
     }
 
-    $scope.selectEid(eid);
+    $scope.selectedId = eid;
   }
 
 
@@ -340,21 +341,29 @@ angular.module('sequoiaGroveApp')
   }
 
   $scope.importLastWeek = function() {
+    $scope.importing = true;
+    $scope.selectedId = 0;
     var d = moment($scope.date.mon.val,'DD-MM-YYYY').subtract(7, 'days').format('DD-MM-YYYY');
-    $scope.getScheduleTemplate(d);
+   $scope.getScheduleTemplate(d)
+     .then(function() {
+       $scope.importing = false;
+   });
   }
 
 /************** HTTP Request Functions **************/
 
   // Save these shift schedulings in the list of updateShifts
   $scope.saveSchedule = function() {
+    if ($scope.saving) {
+      return;
+    }
+    $scope.saving = true;
+    $scope.selectedId = 0;
     // remove blank spaces from update list - they are in delete shifts, or
     // have not been assigned
     $scope.updateShifts = _.filter($scope.updateShifts, function(shift) {
       return (shift.eid !== 0);
     });
-    $log.debug($scope.updateShifts);
-    $scope.saving = true;
 
     $http({
       url: '/sequoiagrove/schedule/update/',
@@ -364,7 +373,6 @@ angular.module('sequoiaGroveApp')
       if (status == 200) {
         // clear update shifts list
         $scope.updateShifts = [];
-        $scope.saving = false;
         $scope.deleteSchedule();
       }
       else {
@@ -378,8 +386,9 @@ angular.module('sequoiaGroveApp')
 
   // Delete these shift schedulings
   $scope.deleteSchedule = function() {
-    $scope.saving = true;
-
+    if ($scope.saving) {
+      return;
+    }
     $http({
       url: '/sequoiagrove/schedule/delete/',
       method: "DELETE",
@@ -412,9 +421,9 @@ angular.module('sequoiaGroveApp')
 
 /************** Event Watchers **************/
 
-  $scope.$watch('selectedId', function(newVal, oldVal){
+  $scope.$watch($rootScope.loading, function(newVal, oldVal){
     if(newVal){
-      $log.debug(newVal);
+      //$log.debug(newVal);
       // watchExpression has changed.
     }
   });
